update screen structure
